check_prices/requirements.txt
anyio==4.11.0
attrs==25.4.0
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
exceptiongroup==1.3.1
Flask==3.0.3
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.1.0
outcome==1.3.0.post0
PySocks==1.7.1
python-dotenv==1.0.1
python-telegram-bot==20.7
requests==2.32.5
selenium==4.21.0
selenium-stealth==1.0.6
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
trio==0.32.0
trio-websocket==0.12.2
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
wsproto==1.3.2
